requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
psycopg2==2.9.6
python-dotenv==1.0.0
marshmallow==3.19.0
requests==2.31.0